+++ apps/fluent-tester/src/FluentTester/TestComponents/Button/ButtonIconTest.tsx
@@ -4,7 +4,7 @@ import { Stack } from '@fluentui-react-native/stack';
 import * as React from 'react';
 import { Platform, View } from 'react-native';
 import { stackStyle } from '../Common/styles';
-import { SvgIconProps, FontIconProps } from '@fluentui-react-native/icon';
+import { SvgIconProps } from '@fluentui-react-native/icon';
 import TestSvg from './test.svg';
 
 const CustomizedIconButton = Button.customize({
@@ -13,48 +13,35 @@ const CustomizedIconButton = Button.customize({
 });
 
 export const ButtonIconTest: React.FunctionComponent<{}> = () => {
-  if (Platform.OS == ('win32' as any)) {
-    const testImage = require('./icon_24x24.png');
-    const testTtf = require('./Font Awesome 5 Free-Solid-900.otf');
+  const testImage = require('./icon_24x24.png');
 
-    const fontProps: FontIconProps = {
-      fontFamily: `Font Awesome 5 Free`,
-      fontSrcFile: testTtf,
-      codepoint: 0xf083,
-      fontSize: 16,
-    };
+  const svgProps: SvgIconProps = {
+    src: TestSvg,
+    viewBox: '0 0 500 500',
+  };
 
-    const svgProps: SvgIconProps = {
-      src: TestSvg,
-      viewBox: '0 0 500 500',
-    };
+  // SVG-based icons are not available on all platforms yet
+  const svgIconsEnabled = ['ios', 'macos', 'win32', 'android'].includes(Platform.OS as string);
 
-    return (
-      <View>
-        <Stack style={stackStyle}>
-          <Button icon={testImage} content="Button with png Icon" tooltip="button tooltip" />
-          <Button icon={{ fontSource: fontProps, color: 'blue' }} content="Button with font Icon" tooltip="button tooltip" />
+  return (
+    <View>
+      <Stack style={stackStyle}>
+        <Button icon={testImage} content="Button with png Icon" tooltip="button tooltip" />
+        {svgIconsEnabled ? (
           <Button
             icon={{ svgSource: svgProps, width: 20, height: 20, color: 'red' }}
             content="Button with svg Icon"
             tooltip="button tooltip"
           />
+        ) : null}
+        {svgIconsEnabled ? (
           <CustomizedIconButton
             icon={{ svgSource: svgProps, width: 20, height: 20 }}
             content="Button with Customized Icon"
             tooltip="button tooltip"
           />
-        </Stack>
-      </View>
-    );
-  } else {
-    return (
-      <View>
-        <Stack style={stackStyle}>
-          <Button icon={require('./icon_24x24.png')} content="Button with Icon" tooltip="button tooltip" />
-          <CustomizedIconButton icon={require('./icon_24x24.png')} content="Button with Customized Icon" tooltip="button tooltip" />
-        </Stack>
-      </View>
-    );
-  }
+        ) : null}
+      </Stack>
+    </View>
+  );
 };